requirements.txt
--index-url https://pypi.devinfra.sentry.io/simple

click>=8.1.7
dictdiffer>=0.9.0
gitpython>=3.1.41
google_api_python_client>=2.137.0
google-cloud-secret-manager>=2.20.2
httpx>=0.27.2
Jinja2>=3.1.4
jsonpatch>=1.33
jsonpath_ng>=1.6.1
jsonschema>=4.23.0
kubernetes>=30.1.0
paramiko>=3.4.0
protobuf>=5.28.1
PyYAML>=6.0.1
referencing>=0.35.1
Requests>=2.32.3
sentry_jsonnet>=0.0.5
sentry_sdk>=2.10.0
setuptools>=69.0.3
tabulate>=0.9.0
